+++ src/index.js
@@ -3,6 +3,7 @@ import './commands/get-user-info';
 import './commands/get-user-tokens';
 import './commands/login';
 import './commands/logout';
+import './commands/preserve-auth0-cookies';
 import './commands/set-auth0-cookie';
 
 import './utils/whitelist-cookies';