+++ CHANGELOG.md
@@ -8,6 +8,8 @@
   `logoutUrl`
 - added new command `cy.clearAuth0Cookies()` to clear all auth0 cookies manually
   anytime
+- added new command `cy.preserveAuth0Cookies()` to preserve cookies between
+  tests
 - added support for customizing cookie settings similar to how `nextjs-auth0`
   does (e.g. with `auth0CookieHttpOnly`). Take a look at the README for more
   details.